apache_airflow/requirements.txt
requests==2.31.0
gspread==6.0.2
apache-airflow==2.9.0
pika==1.3.2
pandas==2.2.2